requirements.txt
alembic==0.9.7
bitcoin == 1.1.42
coincurve == 7.1.0
click==6.7
colour-runner==0.0.5
eth-abi==0.5.0
eth-tester==0.1.0b15
eth-utils==0.7.4
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
flask-socketio==2.9.6
Flask-SQLAlchemy==2.3.2
gevent==1.2.2
gevent-websocket==0.10.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
merkletools == 1.0.3
pyperclip==1.6.0
pycryptodome==3.4.8
python-dateutil==2.6.1
python-editor==1.0.3
repoze.lru == 0.7
rlp==0.6.0
six==1.11.0
SQLAlchemy==1.2.1
unqlite==0.6.2
web3==3.16.4
Werkzeug==0.14.1
